Update MiUnlockTool.py
Hello,

The intention with this update is to make clear that you will need to repeat the process after the time runs out. Since there's no instructions or documentation regarding this process or the time Xiaomi gives to complete the unlock process, it's useful to make things as explicit as possible.

Needless to say, this is just a suggestion.
Thank you.